+++ backend/app/Api/User/Auth/Register.php
@@ -22,6 +22,7 @@ use MythicalDash\CloudFlare\CloudFlareRealIP;
 use MythicalDash\Plugins\Events\Events\AuthEvent;
 use MythicalDash\Chat\IPRelationships\IPRelationship;
 use MythicalDash\Plugins\Events\Events\ReferralsEvent;
+use MythicalDash\Hooks\MythicalSystems\User\UUIDManager;
 use MythicalDash\Hooks\MythicalSystems\CloudFlare\Turnstile;
 use MythicalDash\Services\Pterodactyl\Admin\Resources\UsersResource;